+++ assistant/lib/home/home.dart
@@ -1,7 +1,25 @@
+// Copyright 2024 defsub
+//
+// This file is part of Takeout.
+//
+// Takeout is free software: you can redistribute it and/or modify it under the
+// terms of the GNU Affero General Public License as published by the Free
+// Software Foundation, either version 3 of the License, or (at your option)
+// any later version.
+//
+// Takeout is distributed in the hope that it will be useful, but WITHOUT ANY
+// WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for
+// more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with Takeout.  If not, see <https://www.gnu.org/licenses/>.
+
+import 'package:assistant/settings/repository.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:assistant/settings/repository.dart';
+
 import 'model.dart';
 import 'repository.dart';
 
@@ -25,6 +43,12 @@ class HomeCubit extends Cubit<HomeState> {
       {required this.repository, required this.assistantSettingsRepository})
       : super(HomeState.initial()) {
     _load();
+  }
+
+  @override
+  void emit(HomeState state) {
+    print('emit $state');
+    super.emit(state);
   }
 
   void _load() {