+++ assistant/lib/home/repository.dart
@@ -1,3 +1,20 @@
+// Copyright 2024 defsub
+//
+// This file is part of Takeout.
+//
+// Takeout is free software: you can redistribute it and/or modify it under the
+// terms of the GNU Affero General Public License as published by the Free
+// Software Foundation, either version 3 of the License, or (at your option)
+// any later version.
+//
+// Takeout is distributed in the hope that it will be useful, but WITHOUT ANY
+// WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for
+// more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with Takeout.  If not, see <https://www.gnu.org/licenses/>.
+
 import 'dart:ui';
 
 import 'package:assistant/home/hue_provider.dart';
@@ -13,7 +30,8 @@ class HomeRepository {
 
   Future<bool> authRequired() async => _provider.authRequired();
 
-  Future<void> discover({String? defaultBridge}) async => _provider.discover(defaultBridge: defaultBridge);
+  Future<void> discover({String? defaultBridge}) async =>
+      _provider.discover(defaultBridge: defaultBridge);
 
   Future<void> fetchNetwork() async => _provider.fetchNetwork();